format for Basket, int to double on attribute price for Item

+++ ECommerce.Project/Item.cs
@@ -14,6 +14,6 @@ namespace ECommerce.Project
         public string ItemId;
         public string ItemName;
         public string Description;
-        public int Price;
+        public double Price;
     }
 }
